fixes Github pr_size collector, now as float, causing 500 error

+++ backend/app/services/collectors/models.py
@@ -47,7 +47,7 @@ class GitHubCollectedMetrics(BaseModel):
     pr_review_ratio: float | None
 
     # pr_size
-    pr_size_median: int | None
+    pr_size_median: float | None
 
     # review_turnaround
     review_turnaround_hours: float | None
